pyproject.toml
[project]
name = "atlas"
version = "0.1.0"
description = "Bulding my first AI Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
